\name{pfexamplesinr-package}
\alias{pfexamplesinr-package}
\alias{pfexamplesinr}
\docType{package}
\title{
\packageTitle{pfexamplesinr}
}
\description{
\packageDescription{pfexamplesinr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{pfexamplesinr}
\packageIndices{pfexamplesinr}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{pfexamplesinr}

Maintainer: \packageMaintainer{pfexamplesinr}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
