% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{svolApproxLL}
\alias{svolApproxLL}
\title{Calculates approximate log-likelihood of simple stochastic volatility model (Taylor '82)}
\usage{
svolApproxLL(y, thetaProposal, uProposal)
}
\arguments{
\item{y}{univariate time series vector}

\item{thetaProposal}{parameter vector (order is phi, beta, sigma)}

\item{uProposal}{standard normal variates of dimension time X (particles + 1)}
}
\value{
approximate log-likelihood
}
\description{
Calculates approximate log-likelihood of simple stochastic volatility model (Taylor '82)
}
